requirements.txt
amqp (1.4.6)
anyjson (0.3.3)
argparse (1.2.1)
billiard (3.3.0.19)
celery (3.1.17)
Django (1.7.3)
django-celery (3.1.16)
django-filter (0.9.1)
djangorestframework (3.0.3)
kombu (3.0.24)
Markdown (2.5.2)
pip (1.5.4)
pytz (2014.10)
requests (2.5.1)
setuptools (2.2)
wsgiref (0.1.2)
